requirements.txt
flask
boto3
flask_cors
Flask-JWT-Extendeded